requirements.txt
feedparser
requests
beautifulsoup4
google-generativeai